refactor: move live feed tasks to dedicated scheduler
