+++ src/main/java/se/hydroleaf/scheduler/LiveFeedConfig.java
@@ -1,0 +1,16 @@
+package se.hydroleaf.scheduler;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.time.Instant;
+import java.util.concurrent.ConcurrentHashMap;
+
+@Configuration
+public class LiveFeedConfig {
+
+    @Bean
+    public ConcurrentHashMap<String, Instant> lastSeen() {
+        return new ConcurrentHashMap<>();
+    }
+}
